journal: remove redundant code
Even without this, when we now save our new journal via the web
interface, we get a 'nil' type for the 'avatar' field.

Actually, we ran into a strange bug when this code was in place, where
the 'avatar' field would be assigned the value '_old_' by
carrierwave-mongoid. Removing the code fixes this.