journals: set avatar_type to nil when requested from form
This is an ugly, hacky way to achieve essentially what is a callback on
the model after it is saved. We cannot use an 'after_save' callback on
the Journal model, because 'avatar' does *NOT* belong to the model, and
so the callbacks will never get called when we do a "Remove avatar"
checkbox on the edit form.