journals: put callback in correct sequence
We need to use the 'before_validation' callback, because we need to
validate the 'avatar_type' field with 'validates_with
VldBinaryTypesRange'.

Actually, the validation itself, as far as 'set_avatar_type' is
concerned, is redundant and unnecessary, as the setting of the value is
done purly by the computer and not by the human, on an input dependent
on the human. I.e., 'set_avatar_type' is itself a form of validation.
Still, we leave the validation alone as that can later be expanded to
validate more things.