fix rails console not picking up VldBinaryTypesRange
Follow these steps:
  - make a syntax error in app/models/journal.rb (instead of
    'require_relative', do 'equire_relative') to make it break entirely
  - run 'rails console'
  - try 'Journal.new'; it should obiously break b/c of your purposeful
    syntax error in (1)
  - fix the syntax error
  - reload 'rails console', and do 'Journal.new' again
  - it should complain "NameError: uninitialized constant Journal::VldBinaryTypesRange"
  - in your UNIX shell, run 'touch config/application.rb'
  - reload 'rails console', and do 'Journal.new' again
  - it should now work

The above steps highlight either a bug in Rails core, or my own
library/framework setup bug. The fact that we have to update the
timestamp on 'config/application.rb' indicates that this is indeed a bug
in the library/framework, as solving the problem does not touch a single
bit of code on my end.

Evidently, the rails console is not picking up the validation file, so
the solution is to forcefully add it into 'config/boot.rb'. This
necessitates adding the line "require 'active_model'" in
`validators.rb`, as it is still in the early stage of the booting
process.