models: take care of <mount>_type field outside of model class
The crux of the problem is that the <mount> field of these models are
handled by CarrierWave/GridFs. While the forms work to
create/update these models, the '<mount>_type' field is left alone by
these operations on the model. So, setting the '<mount>_type' from the
controller is far more sensible.