journal: basic rendering framework
We move all rendering details into a helper, as we'll use it to do
on-the-fly rendering on the views through the controller, as well as
saving it later at the user's option. The problem with just rendering
once and saving everything as a string is that if the component
notes/attachments are ever changed, the rendered string is not updated.
This is an optimization problem that can be tackled later on, after
everything else is done.