css: major site facelift
The oldest 'bug' that was fixed was that we now added a simple
'.vert-spacer' div, to get around the fact that using 'position:
relative' for the hand glyphicons in the carousel broke their 'top: 66%'
value in Chromium. By manually adding in a blank element and setting its
height, we can avoid CSS hackery completely.