journal/exhibit: fake the 'back' button
'history.back()' is unreliable --- it works in Firefox but not in Chrome
(and who knows what other browsers break). We instead just link back to
the homepage.

As an aside, this implies that we shouldn't link to a journal exhibit
from another journal exhibit, as this would expose the fakeness of our
implementation here.